// Utils/Language/Enums/ModuleType.ts
export enum ModuleTypes {
    DASHBOARD = "001.MODULE.DASHBOARD",
    EMPLOYEE_MANAGEMENT = "002.MODULE.EMPLOYEE_MANAGEMENT",
    SYSTEM_CONFIGURATION = "003.MODULE.SYSTEM_CONFIGURATION",
    CENTER_MANAGEMENT = "004.MODULE.CENTER_MANAGEMENT",
    FINANCIAL_MANAGEMENT = "005.MODULE.FINANCIAL_MANAGEMENT",
    STUDENT_MANAGEMENT = "006.MODULE.STUDENT_MANAGEMENT",
    TEACHER_MANAGEMENT = "007.MODULE.TEACHER_MANAGEMENT",
    INSPECTION_MANAGEMENT = "008.MODULE.INSPECTION_MANAGEMENT",
    REPORT_MANAGEMENT = "009.MODULE.REPORT_MANAGEMENT",
    USER_MANAGEMENT = "010.MODULE.USER_MANAGEMENT",
    MONITORING_MANAGEMENT = "011.MODULE.MONITORING_MANAGEMENT",
    ASSESSMENT_MANAGEMENT = "012.MODULE.ASSESSMENT_MANAGEMENT",
}