// Utils/Language/Enums/SubChildModuleType.ts
export enum SubChildModuleTypes {
    ADD_EDUCATION_CENTER = "4011.SUBCHILDMODULE.ADD_EDUCATION_CENTER",
    EDUCATION_CENTER_LIST = "4012.SUBCHILDMODULE.EDUCATION_CENTER_LIST",
    ADD_RESOURCE_CENTER = "4021.SUBCHILDMODULE.ADD_RESOURCE_CENTER",
    RESOURCE_CENTER_LIST = "4022.SUBCHILDMODULE.RESOURCE_CENTER_LIST",
    MONTHLY_ASSESSMENT = "6031.SUBCHILDMODULE.MONTHLY_ASSESSMENT",
    MONTHLY_ASSESSMENT_LIST = "6032.SUBCHILDMODULE.MONTHLY_ASSESSMENT_LIST",
    ASSESSMENT_SUMMARY = "6033.SUBCHILDMODULE.ASSESSMENT_SUMMARY",
    YEARLY_ASSESSMENT = "6034.SUBCHILDMODULE.YEARLY_ASSESSMENT",
    YEARLY_ASSESSMENT_LIST = "6035.SUBCHILDMODULE.YEARLY_ASSESSMENT_LIST"
}