// Utils/Language/Enums/SubModuleType.ts
export enum SubModuleTypes {
    ADD_EMPLOYEE = "201.SUBMODULE.ADD_EMPLOYEE",
    EMPLOYEE_LIST = "202.SUBMODULE.EMPLOYEE_LIST",
    DIVISION = "301.SUBMODULE.DIVISION",
    DISTRICT = "302.SUBMODULE.DISTRICT",
    UPAZILA = "303.SUBMODULE.UPAZILA",
    EMPLOYEETYPE = "304.SUBMODULE.EMPLOYEETYPE",
    DESIGNATION = "305.SUBMODULE.DESIGNATION",
    Office = "306.SUBMODULE.OFFICE",
    OFFICETYPE = "307.SUBMODULE.OFFICETYPE",
    LANGUAGE = "308.SUBMODULE.LANGUAGE",
    EDUCATION_CENTER = "401.SUBMODULE.EDUCATION_CENTER",
    RESOURCE_CENTER = "402.SUBMODULE.RESOURCE_CENTER",
    STUDENT_ADD = "601.SUBMODULE.STUDENT_ADD",
    STUDENT_LIST = "602.SUBMODULE.STUDENT_LIST",
    ASSMESSMENT = "603.SUBMODULE.ASSMESSMENT",
    TEACHER_ADD = "701.SUBMODULE.TEACHER_MANAGEMENT",
    TEACHER_LIST = "702.SUBMODULE.TEACHER_MANAGEMENT",
    TEACHER_EVALUATION = "703.SUBMODULE.TEACHER_EVALUATION",
    EVALUATION_SUMMARY = "704.SUBMODULE.EVALUATION_SUMMARY",
    TEACHER_INFO = "705.SUBMODULE.TEACHER_INFO",
}