// Utils/Language/Enums/TitleTypes.ts
export enum TitleTypes {
    TITLE = "TITLE",
    RESET = "RESET",
    BACK = "BACK",
    SEARCH = "SEARCH",
    TYPEHERE = "TYPEHERE",
    SUBMIT = "SUBMIT",
    MORESUBMIT = "MORESUBMIT",
    UPDATE = "UPDATE",
    SL = "SL#",
    LIST = "LIST",
    EDIT = "EDIT",
    ADD = "ADD",
    STATUS = "STATUS",
    ACTION = "ACTION",
    EXPORTTOPDF = "EXPORTTOPDF",
    EXPORTTOEXCEL = "EXPORTTOEXCEL",
    PRINT = "PRINT",
    NAMEEN = "NAMEEN",
    NAMEBN = "NAMEBN",
    PRIORITY = "PRIORITY",
    ADDTITLE = "ADDTITLE",
    EDITTITLE = "EDITTITLE",
    SELECTDIVISION = "SELECTDIVISION",
    SELECTDISTRICT = "SELECTDISTRICT",
    SELECTUPAZILA = "SELECTUPAZILA",
    SELECTOFFICETYPE = "SELECTOFFICETYPE",
    ADDRESS = "ADDRESS",
    OFFICERID = "OFFICERID",
    OFFICERNAMEEN = "OFFICERNAMEEN",
    OFFICERNAMEBN = "OFFICERNAMEBN",
    FATHERSNAME = "FATHERSNAME",
    FATHERSNAMEEN = "FATHERSNAMEEN",
    FATHERSNAMEBN = "FATHERSNAMEBN",
    MOTHEREN = "MOTHEREN",
    MOTHERSNAME = "MOTHERSNAME",
    NID = "NID",
    OFFICERTYPE = "OFFICERTYPE",
    OFFICE = "OFFICE",
    DESIGNATION = "DESIGNATION",
    DATEOFJOINING = "DATEOFJOINING",
    EMAIL = "EMAIL",
    MOBILE = "MOBILE",
    IMAGE = "IMAGE",
    ADDEMPLOYEEBUTTON = "ADDEMPLOYEEBUTTON",
    DIVISION = "DIVISION",
    DISTRICT = "DISTRICT",
    DISTRICTEN = "DISTRICTEN",
    DISTRICTBN = "DISTRICTBN",
    UPAZILA = "UPAZILA",
    CENTERCODE = "CENTERCODE",
    CENTERNAME = "CENTERNAME",
    CITYCORPORATION = "CITYCORPORATION",
    CENTERDESCRIPTION = "CENTERDESCRIPTION",
    TYPEOFCITYCORPORATION = "TYPEOFCITYCORPORATION",
    CENTERPLACE = "CENTERPLACE",
    CENTERPICTURE = "CENTERPICTURE",
    STARTCLASSTIME = "STARTCLASSTIME",
    ENDCLASSTIME = "ENDCLASSTIME",
    CENTERLOCATION = "CENTERLOCATION",
    CENTERADDRESS = "CENTERADDRESS",
    MOSQUECENTERNAME = "MOSQUECENTERNAME",
    STUDENTROLL = "STUDENTROLL",
    STUDENTNAME = "STUDENTNAME",
    GENDER = "GENDER",
    MALE = "MALE",
    FEMALE = "FEMALE",
    OTHER = "OTHER",
    EMPLOYEEID = "EMPLOYEEID",
    CONTACT = "CONTACT",
    OFFICENAME = "OFFICENAME",
    ADDSTUDENTBUTTON = "ADDSTUDENTBUTTON",
    CENTERLEVEL = "CENTERLEVEL",
    CREATEMESSAGE = 'CREATEMESSAGE',
    UPDATEMESSAGE = 'UPDATEMESSAGE',
    DATEOFBIRTH = "DATEOFBIRTH",
    AGE = "AGE",
    BIRTHREGNUMBER = "BIRTHREGNUMBER",
    EDUCATIONYEAR = "EDUCATIONYEAR",
    UNION = "UNION",
    WARDNO = "WARDNO",
    VILLAGE = "VILLAGE",
    INSTITUTIONNAME = "INSTITUTIONNAME",
    STUDENTOCCUPATION = "STUDENTOCCUPATION",
    FATHERNAME = "FATHERNAME",
    OCCUPATION = "OCCUPATION",
    MOTHER = "MOTHER",
    MOTHERBN = "MOTHERBN",
    GUARDIANNAME = "GUARDIANNAME",
    ABSENCEOFFATHER = "ABSENCEOFFATHER",
    TEACHERNAME = "TEACHERNAME",
    ADDTEACHEREVALUATION = "ADDTEACHEREVALUATION",
    CENTERCREATEDAT = "CENTERCREATEDAT",
    TEACHERID = "TEACHERID",
    CONSTITUTIONDATE = "CONSTITUTIONDATE",
    COURSENAME = "COURSENAME",
    SESSION = "SESSION",
    BANKNAME = "BANKNAME",
    BRANCHNAME = "BRANCHNAME",
    ACCOUNTNUMBER = "ACCOUNTNUMBER",
    EDUCATIONQUALIFICATION = "EDUCATIONQUALIFICATION",
    TEACHERSTYPE = "TEACHERSTYPE",
    TRAININGINSTITUTE = "TRAININGINSTITUTE",
    IMAM = "IMAM",
    CERTIFICATE = "CERTIFICATE",
    YEAR = "YEAR",
    BANGLA = "BANGLA",
    ENGLISH = "ENGLISH",
    CENTERTYPE = "CENTERTYPE",
    ADMISSIONDATE = "ADMISSIONDATE",
    ISADMITTED = "ISADMITTED",
    BLOODGROUP = "BLOODGROUP",
    FATHERNID = "FATHERNID",
    MOTHERNID = "MOTHERNID",
    GUARDIANNID = "GUARDIANNID",
    UPDATESTUDENT = "UPDATESTUDENT",
    CENTERNAMEANDCODE = "CENTERNAMEANDCODE",
    AREA = "AREA",
    ROUTING = "ROUTING",
    NATURE = "NATURE",
    TRAINING = "TRAINING",
    NAMEOFDEGREE = "NAMEOFDEGREE",
    DEPARTMENT = "DEPARTMENT",
    BOARDUNIVERSITY = "BOARDUNIVERSITY",
    PASSINGYEAR = "PASSINGYEAR",
    CGPAGRADE = "CGPAGRADE",
    ADDTEACHERDEGREE = "ADDTEACHERDEGREE",
    LEARNINGCENTERID = "LEARNINGCENTERID",
    STARTTIME = "STARTTIME",
    ENDTIME = "ENDTIME",
    EDITTEACHER = "EDITTEACHER",
    UPDATEEMPLOYEE = "UPDATEEMPLOYEE",
    SAVEDRAFT = "SAVEDRAFT",
    MONTH = "MONTH",
    EDUCATIONMATERIAL = "EDUCATIONMATERIAL",
    PHOTO = "PHOTO",
    COURSEID = "COURSEID",
    CENTER = "CENTER",
    ACADEMICQUALIFICATION = "ACADEMICQUALIFICATION",
    INSTITUTENAME = "INSTITUTENAME",
    EMPLOYEENAME = "EMPLOYEENAME",
}